Break SigninView into smaller component
